Update WordCount.py
Changing the regular expression to a simpler form. Expressing the same idea. There is one concern is that in my opinion the line 25 of the code:
Pref.wrdRx                  = re.compile("^\S+$", re.U)
should only be:
Pref.wrdRx                  = re.compile("\S+", re.U)
without the ^ and $ symbols. But since the original codes have them I put it here as a concern.