Changing the regular expression to capture the sentence
Sorry I am new to github so I will try my best to explain myself here. I think the regular expression want to capture a sentence that contains some words. So first it has to be begun with the beginning sentence character **^**, and then it might or might not have a set of any characters including white space character __.*__, but then there must exist a set of non-space character __\S+__ (including __!@#$%^&*()__, and then a set of any characters __.*__ until the end of sentence __$__. Because the string __%^%@#__ is still valid and have words in it, but original regular expression seems to fail to capture it.